# letese/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: LETESE● Legal SaaS API
  description: |
    Legal Practice Management SaaS for Indian law firms.
    Target courts: Punjab & Haryana High Court, Delhi HC, Supreme Court, NCDRC, district courts.
    **Authentication**: All endpoints (except `/auth/*`) require `Authorization: Bearer <access_token>`.
    **RBAC roles**: super_admin, admin, advocate, clerk, paralegal, intern
  version: 1.0.0
  contact:
    name: LETESE Legal Technologies Pvt. Ltd.
    email: api@letese.xyz
  license:
    name: Proprietary
servers:
  - url: https://api.letese.xyz/api/v1
    description: Production
  - url: https://staging-api.letese.xyz/api/v1
    description: Staging
...
paths:
...
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Create tenant + admin user
      description: |
        Creates a new tenant (law firm) and its first admin user in a single call.
        Sends OTP to the admin email for verification.
        **Rate limit**: 3 signups per IP per hour.
      operationId: auth_signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              tenant_name: "Sharma & Associates"
              tenant_email: "contact@sharma-law.in"
              tenant_phone: "+919876543210"
              plan: "professional"
              bar_enrolment_no: "BAR/2012/45678"
              gstin: "09AALFS1234M1ZB"
              admin_email: "rajesh@sharma-law.in"
              admin_full_name: "Rajesh Sharma"
              admin_phone: "+919876543210"
      responses:
        '201':
          description: Tenant and admin user created, OTP sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
        '409':
          description: Tenant or user already exists.
        '429':
          description: Rate limit exceeded.
  /auth/send-otp:
    post:
      tags: [Authentication]
      summary: Send OTP
      description: |
        Sends a 6-digit OTP to the registered email address.
        OTP valid for 10 minutes (600 seconds).
        **Rate limit**: 5 requests per IP per minute.
      operationId: auth_send_otp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "OTP sent" }
                  expires_in: { type: integer, example: 600 }
        '404':
          description: User not found.
        '429':
          description: Rate limit exceeded.
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email + OTP
      description: |
        Verifies OTP, returns JWT access + refresh tokens.
        **Rate limit**: 5 attempts per IP per minute.
      operationId: auth_login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT tokens issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired OTP.
        '429':
          description: Rate limit exceeded.
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      description: Revokes the refresh token. Client should discard the access token.
      operationId: auth_logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Logged out successfully.
        '401':
          description: Invalid or already-revoked refresh token.
  /auth/google:
    post:
      tags: [Authentication]
      summary: Google OAuth login
      description: Login/register via Google OAuth ID token.
      operationId: auth_google
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties:
                id_token:
                  type: string
                  description: Google OAuth ID token from the frontend Google Sign-In flow.
      responses:
        '200':
          description: JWT tokens issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Sign up first before linking Google account.
        '401':
          description: Invalid Google token.
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Exchange a valid refresh token for a new access token.
      operationId: auth_refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, const: "Bearer" }
        '401':
          description: Invalid or expired refresh token.
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Returns the authenticated user's profile and tenant info.
      operationId: auth_me
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
                  tenant:
                    $ref: '#/components/schemas/TenantProfile'
        '401':
          description: Invalid or missing token.

  # ══════════════════════════════════════════════════════════
  # CASES
  # ══════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/google`.
  schemas:
...
    SignupRequest:
      type: object
      required: [tenant_name, tenant_email, tenant_phone, admin_email, admin_full_name]
      properties:
        tenant_name:
          type: string
          minLength: 2
          maxLength: 255
          example: "Sharma & Associates"
        tenant_email:
          type: string
          format: email
          example: "contact@sharma-law.in"
        tenant_phone:
          type: string
          minLength: 10
          maxLength: 20
          example: "+919876543210"
        plan:
          type: string
          enum: [basic, professional, elite, enterprise]
          default: basic
        bar_enrolment_no:
          type: string
          maxLength: 100
          example: "BAR/2012/45678"
        gstin:
          type: string
          maxLength: 20
          example: "09AALFS1234M1ZB"
        firm_address:
          type: string
        admin_email:
          type: string
          format: email
          example: "rajesh@sharma-law.in"
        admin_full_name:
          type: string
          minLength: 2
          maxLength: 255
          example: "Rajesh Sharma"
        admin_phone:
          type: string
          maxLength: 20
    SignupResponse:
      type: object
      properties:
        message: { type: string }
        tenant_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        expires_in: { type: integer, example: 600 }
    SendOtpRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    LoginRequest:
      type: object
      required: [email, otp]
      properties:
        email: { type: string, format: email }
        otp: { type: string, minLength: 6, maxLength: 6, example: "A3F2B9" }
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, const: "Bearer" }
        user:
          $ref: '#/components/schemas/UserObject'
        tenant:
          $ref: '#/components/schemas/TenantObject'
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    UserObject:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role:
          type: string
          enum: [super_admin, admin, advocate, clerk, paralegal, intern]
        tenant_id: { type: string, format: uuid }
    TenantObject:
      type: object
      properties:
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        plan:
          type: string
          enum: [basic, professional, elite, enterprise]
        status: { type: string }
    UserProfile:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role:
          type: string
          enum: [super_admin, admin, advocate, clerk, paralegal, intern]
        phone: { type: string }
        is_active: { type: boolean }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    TenantProfile:
      type: object
      properties:
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        plan:
          type: string
          enum: [basic, professional, elite, enterprise]
        status: { type: string }
        storage_gb: { type: number }

    # ── Cases ──────────────────────────────────────────────
...